% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{annotate_tissues}
\alias{annotate_tissues}
\title{Annotate QTL datasets with metadata}
\usage{
annotate_tissues(dat)
}
\description{
Annotate QTL datasets with metadata
}
\examples{
dat <- annotate_tissues(dat)
}
\keyword{metadata}
