% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{catalogueR-package}
\alias{catalogueR}
\alias{catalogueR-package}
\title{Rapid querying, colocalization, and plotting of summary stats
from the eQTL Catalogue.}
\description{
The functions in \strong{catalogueR} are partly derived from the following
\href{http://htmlpreview.github.io/?https://github.com/eQTL-Catalogue/eQTL-Catalogue-resources/blob/master/scripts/eQTL_API_usecase.html}{eQTL Catalogue tutorial}.
Additional eQTL Catalogue Resources:
\itemize{
\item \href{https://github.com/eQTL-Catalogue/eQTL-Catalogue-resources}{GitHub}
\item \href{https://www.ebi.ac.uk/eqtl/api-docs/}{In-depth API documentation}
\item \href{https://www.ebi.ac.uk/eqtl/}{eQTL Catalogue data portal}
}
}
\details{
\strong{Notes on parallelization}:
There's multiple levels to parallelize on.
You can only choose one level at a time:
\describe{
\item{multithread_qtl=TRUE}{Across QTL datasets.}
\item{multithread_loci=TRUE}{Across loci.}
\item{multithread_tabix=TRUE}{Within tabix files.}
}

You can also get a speedup by using tabix instead of the rest API
Test: For 3 loci, and X QTL datasets:
\describe{
\item{RESTful API}{7.5 minutes.}
\item{Tabix}{27 seconds (*clear winner! ~17x speedup).
That said, if you're only query a small number of specific SNPs
(rather than a large range), the RESTful API can sometimes be faster.}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RajLabMSSM/catalogueR}
  \item Report bugs at \url{https://github.com/RajLabMSSM/catalogueR/issues}
}

}
\author{
\strong{Maintainer}: Brian Schilder \email{brian_schilder@alumni.brown.edu} (\href{https://orcid.org/0000-0001-5949-2191}{ORCID})

Authors:
\itemize{
  \item Jack Humphrey \email{Jack.Humphrey@mssm.edu} (\href{https://orcid.org/0000-0002-6274-6620}{ORCID})
  \item Towfique Raj \email{towfique.raj@mssm.edu} (\href{https://orcid.org/0000-0002-9355-5704}{ORCID})
}

}
