% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coord_input.R
\name{check_coord_input}
\alias{check_coord_input}
\title{Check coordinates input}
\usage{
check_coord_input(gwas_data, chrom, bp_lower, bp_upper)
}
\description{
Make sure there's some way to specify genomic coordinates for tabix query.
}
\keyword{internal}
