% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_maf.R
\name{check_maf}
\alias{check_maf}
\title{Check MAF}
\usage{
check_maf(gwas_shared, eqtl_shared, verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\description{
Check Minor Allele Frequency (MAF) column is 
prepared for running \pkg{coloc}. Must be numeric, <1, >0, 
and not contain NAs.
}
\keyword{internal}
