% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC_corplot.R
\name{COLOC_corplot}
\alias{COLOC_corplot}
\title{Plot summary of coloc tests: heatmap}
\usage{
COLOC_corplot(
  coloc_QTLs,
  coloc_thresh = 0.8,
  label_top_snps = 1,
  facets = Study + System + Tissue ~ Locus.eGene,
  show_plot = TRUE,
  save_dir = tempdir(),
  seed = 2022,
  verbose = TRUE
)
}
\arguments{
\item{coloc_QTLs}{Colocalization results from \link[catalogueR]{COLOC_run}.}

\item{coloc_thresh}{Colocalization Posterior Probability threshold,
 using the formula:
\code{(PP.H3 + PP.H4 >= coloc_thresh) & (PP.H4 / PP.H3 >= 2)}.}

\item{label_top_snps}{Label the top n SNPs per \code{facets} group, sorted
by \code{SNP.PP.H4} from \pkg{coloc}.}

\item{facets}{Formula specifying the faceting variables for the plot.
Passed to \link[ggplot2]{facet_grid}.}

\item{show_plot}{Whether to print the plot.}

\item{save_dir}{Directory to save plot in.}

\item{seed}{Random seed for reproducibility of label placement
via \link[ggrepel]{geom_label_repel}.}

\item{verbose}{Print messages.}
}
\description{
Plots the output from \link[catalogueR]{COLOC_run}.
}
\examples{
\dontrun{
coloc_QTLs <- catalogueR::COLOC_get_example_res()
gg_coloc <- COLOC_corplot(coloc_QTLs = coloc_QTLs,
                          coloc_thresh = .5)
}
}
\seealso{
Other coloc: 
\code{\link{COLOC_get_example_res}()},
\code{\link{COLOC_get_res}()},
\code{\link{COLOC_heatmap}()},
\code{\link{COLOC_merge_res}()},
\code{\link{COLOC_report_summary}()},
\code{\link{COLOC_run}()}
}
\concept{coloc}
