% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC_merge_res.R
\name{COLOC_merge_res}
\alias{COLOC_merge_res}
\title{Prepare data for coloc plot}
\usage{
COLOC_merge_res(
  gwas.qtl_paths,
  coloc_QTLs = NULL,
  qtl_thresh = 1e-05,
  coloc_thresh = 0.8,
  gwas_label = "GWAS",
  remove_extra_panes = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gwas.qtl_paths}{Query results paths from 
\link[catalogueR]{eQTLcatalogue_query}.}

\item{coloc_QTLs}{Colocalization results from \link[catalogueR]{COLOC_run}.}

\item{qtl_thresh}{QTL uncorrected p-value ("pvalues.QTL") threshold.}

\item{coloc_thresh}{Colocalization Posterior Probability threshold,
 using the formula:
\code{(PP.H3 + PP.H4 >= coloc_thresh) & (PP.H4 / PP.H3 >= 2)}.}

\item{gwas_label}{Label for the GWAS subplot.}

\item{remove_extra_panes}{Remove SNPs from non-significant panes.}

\item{verbose}{Print messages.}
}
\description{
Use the coloc results (coloc_QTLs) to determine which full summary stats
 (gwas.qtl_paths) to plot.
}
\seealso{
Other coloc: 
\code{\link{COLOC_corplot}()},
\code{\link{COLOC_get_example_res}()},
\code{\link{COLOC_get_res}()},
\code{\link{COLOC_heatmap}()},
\code{\link{COLOC_report_summary}()},
\code{\link{COLOC_run}()}
}
\concept{coloc}
