% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC_report_summary.R
\name{COLOC_report_summary}
\alias{COLOC_report_summary}
\title{Report coloc results}
\usage{
COLOC_report_summary(coloc.res, coloc_thresh = 0.8)
}
\description{
Report coloc results
}
\seealso{
Other coloc: 
\code{\link{COLOC_corplot}()},
\code{\link{COLOC_get_example_res}()},
\code{\link{COLOC_get_res}()},
\code{\link{COLOC_heatmap}()},
\code{\link{COLOC_merge_res}()},
\code{\link{COLOC_run}()}
}
\concept{coloc}
\keyword{internal}
