% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDT.R
\name{createDT}
\alias{createDT}
\title{Create an interactive data table with download buttons}
\usage{
createDT(DF, caption = "", scrollY = 400)
}
\arguments{
\item{DF}{Data table.}

\item{caption}{Table caption.}

\item{scrollY}{Max height in pixels before scrolling.}
}
\description{
Create an interactive data table with download buttons
}
