% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_Catalogue.fetch.R
\name{eQTL_Catalogue.fetch}
\alias{eQTL_Catalogue.fetch}
\title{Query eQTL Catalogue}
\usage{
eQTL_Catalogue.fetch(
  unique_id,
  method = c("REST", "tabix", "echotabix"),
  quant_method = "ge",
  infer_region = TRUE,
  gwas_data = NULL,
  nThread = 1,
  chrom = NULL,
  bp_lower = NULL,
  bp_upper = NULL,
  multithread_tabix = FALSE,
  add_qtl_id = TRUE,
  convert_genes = TRUE,
  conda_env = "echoR",
  verbose = TRUE
)
}
\arguments{
\item{unique_id}{Unique eQTL Catalogue ID assigned in metadata
("unique_id" column in \code{data(meta)}).}

\item{method}{Method for querying eQTL Catalogue:
\itemize{
\item{"tabix"}{Uses tabix CLI tool. Fastest option, but less reliable. 
Requires tabix CLI tool to be installed by user first.}
\item{"echotabix"}{Uses tabix CLI tool. 2nd fastest option,
 but less reliable. 
 Does not require any additional software to be installed.}
\item{"REST"}{Uses the REST API.}{Slowest, but most reliable.
Does not require any additional software to be installed.}
}
\emph{Note}:Tabix is about ~17x faster than the REST API,
 but is currently a far less reliable method than the REST API because 
 tabix tends to get blocked by eQTL Catalogue's firewall.
See \href{https://github.com/RajLabMSSM/catalogueR/issues/5}{here}
for more details.}

\item{quant_method}{eQTL Catalogue actually contains more than just
 eQTL data.
For each dataset, the following kinds of QTLs can be queried:
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default})
or \code{quant_method="microarray"}, depending on the dataset.
\strong{catalogueR} will automatically select whichever option is available.}
\item{exon expression QTL}{\emph{*under construction*}
\code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}
\code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{
\emph{*under construction*}  \code{quant_method="txrev"}}
}}

\item{infer_region}{Infer the range of coordinates to query based on
the min/max positions in each of the summary statistics files
 (\code{sumstats_paths}).}

\item{gwas_data}{\link[data.table]{data.table} of GWAS summary statistics.}

\item{nThread}{The number of CPU cores you want to use to speed up your
queries through parallelization.}

\item{chrom}{Chromosome of the query window.}

\item{bp_lower}{Minimum basepair position of the query window.}

\item{bp_upper}{Maxmimum basepair position of the query window.}

\item{multithread_tabix}{Multi-thread across within a single tabix file query
(good when you have one-several large loci).}

\item{add_qtl_id}{Add "qtl_id" (i.e. "unique_id") column to the query result.}

\item{convert_genes}{Convert Ensembl IDs to HGNC symbols.}

\item{conda_env}{Conda environment to search for tabix executable in.}

\item{verbose}{Show more (\code{=TRUE}) or fewer (\code{=FALSE}) messages.}
}
\description{
Query eQTL Catalogue datasets with multiple methods options.
}
\examples{
data("meta")
GWAS.QTL_manual <- catalogueR:: eQTL_Catalogue.fetch(
    unique_id = meta$unique_id[1],  
    chrom = 8, 
    bp_lower = 21527069-500,
    bp_upper = 21527069+500)
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTL_Catalogue.header}},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
