% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eQTL_Catalogue.header}
\alias{eQTL_Catalogue.header}
\title{eQTL Catalogue tabix header}
\format{
An object of class \code{character} of length 19.
}
\source{
\code{
eQTL_Catalogue.header <- tabix_header(force_new_header = TRUE)
usethis::use_data(eQTL_Catalogue.header, overwrite = TRUE)
}
}
\usage{
data("eQTL_Catalogue.header")
}
\description{
The eQTL Catalogue tabix header (column names)
 is not always included in the file.
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{datasets}
