% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{eQTL_Catalogue.iterate_fetch}
\alias{eQTL_Catalogue.iterate_fetch}
\title{Iterate queries to \emph{eQTL Catalogue}}
\usage{
eQTL_Catalogue.iterate_fetch(
  sumstats_paths,
  output_dir = "./catalogueR_queries",
  qtl_id,
  quant_method = "ge",
  infer_region = T,
  use_tabix = T,
  multithread_loci = T,
  multithread_tabix = F,
  nThread = 4,
  split_files = T,
  merge_with_gwas = F,
  force_new_subset = F,
  progress_bar = F,
  genome_build = "hg19",
  verbose = T
)
}
\arguments{
\item{sumstats_paths}{A list of paths to any number of summary stats files 
whose coordinates you want to use to make queries to eQTL Catalogue.  
If you wish to add custom names to the loci, simply add these as the names of the path list
 (e.g. \code{c(BST1="<path>/<to>/<BST1_file>", LRRK2="<path>/<to>/<LRRK2_file>")}). 
 Otherwise, loci will automatically named based on their min/max genomic coordinates. 
 
The minimum columns in these files required to make queries include: 
\describe{
\item{SNP}{RSID of each SNP.} 
\item{CHR}{Chromosome (can be in "chr12" or "12" format).}  
\item{POS}{Genomic position of each SNP.}  
\item{...}{Optional extra columns.}  
}}

\item{output_dir}{The folder you want the merged gwas/qtl results to be saved to 
(set \code{output_dir=F} if you don't want to save the results).  
If \code{split_files=F}, all query results will be merged into one and saved as \emph{<output_dir>/eQTL_Catalogue.tsv.gz}.  
If \code{split_files=T}, all query results will instead be split into smaller files and stored in \emph{<output_dir>/}.}

\item{quant_method}{eQTL Catalogue actually contains more than just eQTL data.
For each dataset, the following kinds of QTLs can be queried:  
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default}) or \code{quant_method="microarray"}, depending on the dataset. \strong{catalogueR} will automatically select whichever option is available.}  
\item{exon expression QTL}{\emph{*under construction*}  \code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}  \code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{\emph{*under construction*}  \code{quant_method="txrev"}}   
}}

\item{use_tabix}{Tabix is about ~17x faster (\emph{default:} =T) than the REST API (\emph{=F}).}

\item{nThread}{The number of CPU cores you want to use to speed up your queries through parallelization.}

\item{split_files}{Save the results as one file per QTL dataset (with all loci within each file).
If this is set to \code{=T}, then this function will return the list of paths where these files were saved.
A helper function is provided to import and merge them back together in R.  
If this is set to  \code{=F}, then this function will instead return one big merged data.table  
containing results from all QTL datasets and all loci.  
\code{=F} is not recommended when you have many large loci and/or many QTL datasets,  
because you can only fit so much data into memory.}

\item{merge_with_gwas}{Whether you want to merge your QTL query results with your GWAS data 
(convenient, but takes up more storage).}

\item{force_new_subset}{By default, \strong{catalogueR} will use any pre-existing files that match your query. 
Set \code{force_new_subset=T} to override this and force a new query.}

\item{progress_bar}{Show progress bar during parallelization across loci.
\emph{WARNING!}: Progress bar (via \code{\link{pbmclapply}}) only works
 on Linux/Unix systems (e.g. mac) and NOT on Windows.}

\item{genome_build}{The genome build of your query coordinates (e.g. \code{gwas_data}). 
If your coordinates are in \emph{hg19}, \strong{catalogueR} will automatically lift them over 
to \emph{hg38} (as this is the build that eQTL Catalogue uses).}

\item{verbose}{Show more (\code{=T}) or fewer (\code{=F}) messages.}
}
\description{
Uses coordinates from stored summary stats files (e.g. GWAS) 
to determine which regions to query from \emph{eQTL Catalogue}.
}
\examples{
sumstats_paths <- example_sumstats_paths()
qtl_id  <- eQTL_Catalogue.list_datasets()$unique_id[1]
GWAS.QTL <- eQTL_Catalogue.iterate_fetch(sumstats_paths=sumstats_paths, qtl_id=qtl_id, force_new_subset=T, multithread_loci=T, nThread=1, split_files=F, progress_bar=F)
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{LRRK2__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
