% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_Catalogue.search_metadata.R
\name{eQTL_Catalogue.search_metadata}
\alias{eQTL_Catalogue.search_metadata}
\title{Search eQTL Catalogue metadata}
\usage{
eQTL_Catalogue.search_metadata(qtl_search = NULL, verbose = TRUE)
}
\description{
Searches through multiple relevant metadata columns to find eQTL Catalogue datasets
that match at least one of your substrings in a list.
All searches are case-insensitive.
If \code{qtl_search=NULL}, will return all available datasets.
}
\examples{
qtl_datasets <- eQTL_Catalogue.search_metadata(qtl_search = c("Alasoo_2018", "monocyte"))
qtl_datasets.brain <- eQTL_Catalogue.search_metadata(qtl_search = "brain")
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{LRRK2__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.header}},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
