% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eQTLcatalogue_header}
\alias{eQTLcatalogue_header}
\title{eQTL Catalogue tabix header}
\format{
An object of class \code{character} of length 19.
}
\source{
\code{
eQTLcatalogue_header <- eQTLcatalogue_get_header(force_new_header = TRUE)
usethis::use_data(eQTLcatalogue_header, overwrite = TRUE)
}
}
\usage{
data("eQTLcatalogue_header")
}
\description{
The eQTL Catalogue tabix header (column names)
 is not always included in the file.
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTLcatalogue_fetch}()},
\code{\link{eQTLcatalogue_iterate_fetch}()},
\code{\link{eQTLcatalogue_query}()},
\code{\link{eQTLcatalogue_search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{datasets}
