% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLcatalogue_list_datasets.R
\name{eQTLcatalogue_list_datasets}
\alias{eQTLcatalogue_list_datasets}
\title{List available eQTL datasets}
\source{
\href{https://github.com/eQTL-Catalogue/eQTL-Catalogue-resources}{
eQTL-Catalogue GitHub repo}
}
\usage{
eQTLcatalogue_list_datasets(
  save_dir = tempdir(),
  force_new = FALSE,
  include_imported = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{save_dir}{Where to save the processed metadata.}

\item{force_new}{Re-import the metadata from GitHub instead of using
metadata that comes included with \pkg{catalogueR}.}

\item{include_imported}{Include metadata from datasets that have not yet been
fully re-processed by eQTL Catalogue's standardized pipeline.}

\item{verbose}{Print messages.}
}
\description{
Does some additional preprocessing of metadata to categorize tissue types.
}
\examples{
\dontrun{
meta <- catalogueR::eQTLcatalogue_list_datasets()
}
}
\keyword{metadata}
