% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_eQTL_Catalogue_query_paths.R
\name{example_eQTL_Catalogue_query_paths}
\alias{example_eQTL_Catalogue_query_paths}
\title{Paths to example eQTL Catalogue query results}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
example_eQTL_Catalogue_query_paths(Rlib_path = NULL)
}
\arguments{
\item{Rlib_path}{This function will automatically find your Rlib path,
but you can override this by supplying it manually.}
}
\description{
Returns the paths to eQTL Catalogue query results stored within \emph{catalogueR}.
Each file is a merged data.table of the GWAS summary stats used to make the query,
and the eQTL Catalogue query results (which can contain data for multiple eGenes).
}
\details{
GWAS data originally comes from the Parkinson's disease GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.

\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic positiion (in basepairs)}
  ...
}
}
\examples{
gwas.qtl_paths <- example_eQTL_Catalogue_query_paths()
}
\concept{Nalls23andMe_2019
These example files can be used}
