% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sumstats_paths.R
\name{example_sumstats_paths}
\alias{example_sumstats_paths}
\title{Paths to example summary stats}
\source{
\url{https://doi.org/10.1016/S1474-4422(19)30320-5}
}
\usage{
example_sumstats_paths(
  save_dir = file.path(tempdir(), "GWAS/Nalls2019"),
  verbose = FALSE
)
}
\arguments{
\item{save_dir}{Folder to save example summary statistics in.}

\item{verbose}{Print messages.}
}
\description{
Returns the paths to summary stats stored within \emph{catalogueR}.
Each file is the output of a locus that has been fine-mapped
using \link[echolocatoR]{finemap_loci}.
Data originally comes from the Parkinson's disease GWAS
by \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al., 2019 (The Lancet Neurology)}.
}
\details{
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic positiion (in basepairs)}
  ...
}
}
\examples{
sumstats_paths <- catalogueR::example_sumstats_paths()
}
\seealso{
Other Nalls23andMe_2019: 
\code{\link{example_eQTL_Catalogue_query_paths}()}
}
\concept{Nalls23andMe_2019}
