% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_from_eqtl_cat_API.R
\name{fetch_from_eqtl_cat_API}
\alias{fetch_from_eqtl_cat_API}
\title{Fetch from eQTL Catalogue API}
\usage{
fetch_from_eqtl_cat_API(link, is_gwas = FALSE, timeout = 5 * 60)
}
\arguments{
\item{link}{File link.}

\item{is_gwas}{Whether the file is a GWAS dataset.}
}
\description{
Sub-function of fetch_restAPI.
}
\keyword{internal}
