% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tabix.R
\name{fetch_tabix}
\alias{fetch_tabix}
\title{Query eQTL Catalogue:tabix/echotabix}
\source{
\href{https://github.com/RajLabMSSM/catalogueR/issues/5}{
eQTL Catalogue blocking tabix requests}

\code{
data("meta");
query_granges <- echotabix::construct_query(query_dat = echodata::BST1)
qtl.subset <- catalogueR:::fetch_tabix(unique_label=meta$unique_label[2],
                                       query_granges=query_granges)
}
}
\usage{
fetch_tabix(
  unique_label,
  query_granges,
  query_method = c("rsamtools", "variantannotation", "conda", "seqminer"),
  quant_method = "ge",
  nThread = 1,
  conda_env = "echoR_mini",
  verbose = TRUE
)
}
\arguments{
\item{query_granges}{\link[GenomicRanges]{GRanges} object 
to be used for querying the \code{target_path} file. 
Alternatively, can be variant-level summary statistics to be converted into a
\link[GenomicRanges]{GRanges} object by \link[echotabix]{construct_query}.}

\item{query_method}{Method used for querying. 
See \link[echotabix]{query} for available options.}

\item{quant_method}{eQTL Catalogue actually contains more than just
 eQTL data.
For each dataset, the following kinds of QTLs can be queried:
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default})
or \code{quant_method="microarray"}, depending on the dataset.
\strong{catalogueR} will automatically select whichever option is available.}
\item{exon expression QTL}{\emph{*under construction*}
\code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}
\code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{
\emph{*under construction*}  \code{quant_method="txrev"}}
}}

\item{nThread}{The number of CPU cores you want to use to speed up your
queries through parallelization.}

\item{conda_env}{Conda environment to search for tabix executable in.}

\item{verbose}{Show more (\code{=TRUE}) or fewer (\code{=FALSE}) messages.}
}
\description{
Query eQTL Catalogue datasets by region with tabix or \pkg{echotabix}.
Faster alternative to REST API.
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTLcatalogue_fetch}()},
\code{\link{eQTLcatalogue_header}},
\code{\link{eQTLcatalogue_iterate_fetch}()},
\code{\link{eQTLcatalogue_query}()},
\code{\link{eQTLcatalogue_search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{internal}
