% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_consensus_SNPs.R
\name{find_consensus_SNPs}
\alias{find_consensus_SNPs}
\title{Find Consensus SNPs in \emph{echolocatoR} output}
\usage{
find_consensus_SNPs(
  finemap_dat,
  verbose = TRUE,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  sort_by_support = TRUE,
  exclude_methods = NULL
)
}
\description{
Find Consensus SNPs in \emph{echolocatoR} output
}
\examples{
data("BST1")
BST1 <- find_consensus_SNPs(finemap_dat = BST1)
}
\concept{echolocatoR}
