% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gather_colocalized_data}
\alias{gather_colocalized_data}
\title{Prepare data for coloc plot}
\usage{
gather_colocalized_data(
  gwas.qtl_paths,
  coloc_QTLs = NULL,
  qtl_thresh = 1e-05,
  coloc_thresh = 0.8,
  gwas_label = "GWAS",
  remove_extra_panes = T,
  verbose = T
)
}
\description{
Use the coloc results (coloc_QTLs) to determine which full summary stats (gwas.qtl_paths) to plot.
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{coloc_QTLs_full}},
\code{\link{coloc_QTLs}},
\code{\link{get_colocs}()},
\code{\link{plot_coloc_summary}()},
\code{\link{retrieve_sumstats_info}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
