% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_files.R
\name{gather_files}
\alias{gather_files}
\title{Merge files from a list of paths}
\usage{
gather_files(file_paths, nThread = 1, verbose = TRUE)
}
\description{
Merge a list of files into one by stacking them on top of each other (i.e. rbind).
}
\examples{
sumstats_paths <- example_sumstats_paths()
merged_dat <- gather_files(file_paths = sumstats_paths)
}
\seealso{
Other utils: 
\code{\link{ensembl_to_hgnc}()},
\code{\link{hgnc_to_ensembl}()},
\code{\link{liftover}()}
}
\concept{utils}
