% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data via \pkg{piggyback}}
\usage{
get_data(
  fname,
  repo = "RajLabMSSM/catalogueR",
  storage_dir = tempdir(),
  overwrite = FALSE
)
}
\arguments{
\item{fname}{File name.}

\item{repo}{GitHub repository \<owner\>/\<name\>.}

\item{storage_dir}{Data storage directory.}

\item{overwrite}{Overwrite previously downloaded data.}
}
\description{
Download data stored in GitHub Releases using \pkg{piggyback}.
}
\keyword{internal}
