% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Infer (effective) sample size from summary stats}
\usage{
get_sample_size(subset_DT, sample_size = NULL, effective_ss = T, verbose = T)
}
\description{
Infer (effective) sample size from summary stats
}
\examples{
data("BST1")
BST1 <- finemap_DT
subset_DT <- get_sample_size(subset_DT = finemap_DT)
}
\concept{general}
\keyword{internal}
