% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{liftover}
\alias{liftover}
\title{Lift genome across builds}
\usage{
liftover(gwas_data, build.conversion = "hg19.to.hg38", verbose = TRUE)
}
\arguments{
\item{build_conversion}{"hg19.to.hg38" (\emph{default}) or "hg38.to.hg19.}
}
\description{
Lift genome across builds
}
\examples{
data("BST1")
gr.lifted <- liftover(gwas_data = BST1, build.conversion = "hg19.to.hg38")
}
\seealso{
Other utils: 
\code{\link{ensembl_to_hgnc}()},
\code{\link{gather_files}()},
\code{\link{hgnc_to_ensembl}()}
}
\concept{utils}
