% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_plot.R
\name{locus_plot}
\alias{locus_plot}
\title{Locus plot}
\usage{
locus_plot(
  coloc_QTLs,
  locus_name,
  LD_reference = "1KGphase3",
  LD_results_dir = "data/COVID19_GWAS",
  plot_results_dir = "plots",
  force_new_LD = F,
  plot_zoom = "1x",
  plot_title = "eQTL Catalogue",
  top_title = "GWAS",
  top_x_lab = NULL,
  show_plot = TRUE,
  save_plot = TRUE,
  dpi = 350,
  height = 12,
  width = 7
)
}
\description{
Plot colocalized GWAS and QTL results.
}
