% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRRK2__Alasoo_2018.macrophage_IFNg}
\alias{LRRK2__Alasoo_2018.macrophage_IFNg}
\title{eQTL Catalogue query results}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2610 rows and 50 columns.
}
\source{
\code{
root_dir <- "~/Desktop/catalogueR_queries/Alasoo_2018.macrophage_IFNg"
file_name <- "LRRK2__Alasoo_2018.macrophage_IFNg.tsv.gz"
LRRK2__Alasoo_2018.macrophage_IFNg <- data.table::fread(file.path(root_dir,
                                                                  file_name))
usethis::use_data(LRRK2__Alasoo_2018.macrophage_IFNg, overwrite = TRUE)
}
}
\usage{
LRRK2__Alasoo_2018.macrophage_IFNg
}
\description{
Query: LRRK2 GWAS locus, Alasoo_2018.macrophage_IFNg QTL dataset.
Split output file from \link[catalogueR]{eQTL_Catalogue.query}.
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.header}},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{datasets}
