% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gwas_qtl.R
\name{merge_gwas_qtl}
\alias{merge_gwas_qtl}
\title{Merge GWAS with QTL}
\usage{
merge_gwas_qtl(query_dat, qtl.subset, verbose = TRUE)
}
\arguments{
\item{verbose}{Show more (\code{=TRUE}) or fewer (\code{=FALSE}) messages.}
}
\description{
Merge GWAS data (query) and QTL data (results).
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTLcatalogue_fetch}()},
\code{\link{eQTLcatalogue_header}},
\code{\link{eQTLcatalogue_iterate_fetch}()},
\code{\link{eQTLcatalogue_query}()},
\code{\link{eQTLcatalogue_search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{internal}
