% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{eQTL Catalogue dataset metadata}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 758 rows and 14 columns.
}
\source{
\code{
meta <- eQTLcatalogue_list_datasets(force_new = TRUE) 
usethis::use_data(meta, overwrite = TRUE)
}
}
\usage{
meta
}
\description{
List of all queryable tabix-indexed eQTL Catalogue datasets
and their associated systems/tissues/cell types.
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTLcatalogue_fetch}()},
\code{\link{eQTLcatalogue_header}},
\code{\link{eQTLcatalogue_iterate_fetch}()},
\code{\link{eQTLcatalogue_query}()},
\code{\link{eQTLcatalogue_search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()}
}
\concept{eQTL Catalogue}
\keyword{datasets}
