% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coloc_summary.R
\name{plot_coloc_summary}
\alias{plot_coloc_summary}
\title{Plot summary of coloc tests}
\usage{
plot_coloc_summary(
  coloc_QTLs,
  coloc_thresh = 0.8,
  label_snp_groups = FALSE,
  save_dir = tempdir(),
  show_plot = TRUE
)
}
\arguments{
\item{coloc_QTLs}{Colocalization results from \link[catalogueR]{run_coloc}.}

\item{coloc_thresh}{Colocalization Posterior Probability threshold,
 using the formula:
\code{(PP.H3 + PP.H4 >= coloc_thresh) & (PP.H4 / PP.H3 >= 2}.}

\item{label_snp_groups}{Whether to label SNP groups.}

\item{save_dir}{Directory to save plot in.}

\item{show_plot}{Whether to print the plot.}
}
\description{
Plots the output from \link[catalogueR]{run_coloc}.
}
\examples{
coloc_QTLs <- catalogueR::get_coloc_QTLs()
gg_coloc <- catalogueR::plot_coloc_summary(
    coloc_QTLs = coloc_QTLs,
    coloc_thresh = .5
)
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{get_coloc_QTLs}()},
\code{\link{get_colocs}()},
\code{\link{merge_colocalized_data}()},
\code{\link{retrieve_sumstats_info}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
