% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare plot data}
\usage{
prepare_plot_data(
  coloc_QTLs,
  locus_name,
  QTL_gene = NULL,
  GWAS.QTL = NULL,
  LD_reference = "1KGphase3",
  LD_results_dir = file.path(tempdir(), "plot_data"),
  force_new_LD = FALSE,
  verbose = TRUE
)
}
\description{
Prepare plot data.
}
\keyword{internal}
