% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_sumstats_info.R
\name{retrieve_sumstats_info}
\alias{retrieve_sumstats_info}
\title{Retrieve summary statistics info}
\usage{
retrieve_sumstats_info(
  output_dir = file.path(tempdir(), "catalogueR_queries"),
  coloc_QTLs
)
}
\description{
Merge coloc results with SNP-wise GWAS/QTL data
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{get_coloc_QTLs}()},
\code{\link{get_colocs}()},
\code{\link{merge_colocalized_data}()},
\code{\link{plot_coloc_summary}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
\keyword{internal}
