% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsids_from_coords.R
\name{rsids_from_coords}
\alias{rsids_from_coords}
\title{Get RSIDS from SNP coordinates}
\usage{
rsids_from_coords(
  dat,
  genome_build = "hg19",
  drop_unannotated = TRUE,
  drop_duplicates = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data.table of SNP positions,
with the columns \code{CHR} and \code{POS}. Extra columns allowed.}

\item{genome_build}{Which genome build \code{dat} is in
(e.g. "hg19" or "hg38").}

\item{drop_unannotated}{Drop SNPs that RSIDs couldn't be found for.}

\item{drop_duplicates}{Drop any duplicate SNPs rows.}
}
\description{
Get RSIDS from SNP coordinates
}
\examples{
dat <- catalogueR::BST1
dat.annot <- rsids_from_coords(dat, genome_build = "hg19")
}
