% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coloc.R
\name{run_coloc}
\alias{run_coloc}
\title{Iteratively run coloc on merged GWAS-QTL datatables}
\usage{
run_coloc(
  gwas.qtl_paths,
  save_path = file.path(tempdir(), "coloc_results.tsv.gz"),
  nThread = 1,
  top_snp_only = TRUE,
  split_by_group = FALSE,
  method = "abf",
  coloc_thresh = 0.8,
  gwas_sample_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gwas.qtl_paths}{Query results paths from 
\link[catalogueR]{eQTL_Catalogue.query}.}

\item{save_path}{Where to save results to.}

\item{nThread}{The number of CPU cores you want to use to speed up your
queries through parallelization.}

\item{top_snp_only}{Only include the SNP (with the highest SNP-wise PP.H4, 
which is usually the one with the smallest p-value) instead of all SNPs. 
Can be useful for reducing data size.}

\item{split_by_group}{Split files by QTL group when saving.}

\item{method}{Method for querying eQTL Catalogue:
\itemize{
\item{"tabix"}{Uses tabix CLI tool. Fastest option, but less reliable. 
Requires tabix CLI tool to be installed by user first.}
\item{"echotabix"}{Uses tabix CLI tool. 2nd fastest option,
 but less reliable. 
 Does not require any additional software to be installed.}
\item{"REST"}{Uses the REST API.}{Slowest, but most reliable.
Does not require any additional software to be installed.}
}
\emph{Note}:Tabix is about ~17x faster than the REST API,
 but is currently a far less reliable method than the REST API because 
 tabix tends to get blocked by eQTL Catalogue's firewall.
See \href{https://github.com/RajLabMSSM/catalogueR/issues/5}{here}
for more details.}

\item{coloc_thresh}{Colocalization Posterior Probability threshold,
 using the formula:
\code{(PP.H3 + PP.H4 >= coloc_thresh) & (PP.H4 / PP.H3 >= 2}.}

\item{gwas_sample_size}{Sample size of the GWAS data 
(useful if "N" column not already in data).}

\item{verbose}{Print messages.}
}
\value{
If \code{top_snp_only=TRUE}, returns SNP-level stats for only the SNP
with the highest colocalization probability (\emph{SNP.PP.H4})
If \code{top_snp_only=FALSE}, returns SNP-level stats for every SNP.
In either case, summary-level coloc stats are added in the columns
\emph{PP.H0}, \emph{PP.H1}, \emph{PP.H2}, \emph{PP.H3}, \emph{PP.H4}.
}
\description{
Runs colocalization tests (\link[coloc]{coloc.abf})
on merged GWAS-QTL \emph{data.tables}
generated by \link[catalogueR]{eQTL_Catalogue.query}.
Iteratively runs coloc across each:
\itemize{
\item{QTL dataset}
\item{GWAS locus}
\item{QTL gene}
}
\emph{NOTE:} Assumes that each file is within a subfolder named after
 the QTL dataset it came from.
}
\examples{
gwas.qtl_paths <- catalogueR::example_eQTL_Catalogue_query_paths()
coloc_QTLs <- catalogueR::run_coloc(gwas.qtl_paths = gwas.qtl_paths)
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{get_coloc_QTLs}()},
\code{\link{get_colocs}()},
\code{\link{merge_colocalized_data}()},
\code{\link{plot_coloc_summary}()},
\code{\link{retrieve_sumstats_info}()}
}
\concept{coloc}
