% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix_reader.R
\name{tabix_reader}
\alias{tabix_reader}
\title{Read a tabix file subset}
\usage{
tabix_reader(
  tabix_path,
  region,
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE
)
}
\description{
Query a subset of a (remote) tabix file and direclty import it with 
\link[data.table]{fread}.
}
\keyword{internal}
