% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC.report_summary.R
\name{COLOC.report_summary}
\alias{COLOC.report_summary}
\title{Report coloc results}
\usage{
COLOC.report_summary(coloc.res, coloc_thresh = 0.8)
}
\description{
Report coloc results
}
\seealso{
Other coloc: 
\code{\link{get_coloc_QTLs}()},
\code{\link{get_colocs}()},
\code{\link{merge_colocalized_data}()},
\code{\link{plot_coloc_summary}()},
\code{\link{retrieve_sumstats_info}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
\keyword{internal}
