% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC_get_example_res.R
\name{COLOC_get_example_res}
\alias{COLOC_get_example_res}
\title{Example colocalization results}
\source{
\code{
##### Subset results ####
# Pre-processing
gwas.qtl_paths <- eQTLcatalogue_example_queries()
coloc_QTLs <- COLOC_run(gwas.qtl_paths = gwas.qtl_paths,
                        nThread = 4,
                        top_snp_only = TRUE,
                        save_path = "~/Desktop/coloc_results.tsv.gz")
# Import pre-processed results
URL <- file.path("https://github.com/RajLabMSSM/catalogueR/raw/master",
                 "data/coloc_QTLs.rda")
tmp <- file.path(tempdir(),basename(URL))
download.file(URL, tmp)
piggyback::pb_upload(file = tmp,
                     repo = "RajLabMSSM/catalogueR")


##### Full results ####
URL <- file.path("https://github.com/RajLabMSSM/catalogueR/raw/master",
                 "data/coloc_QTLs_full.rda")
tmp <- file.path(tempdir(),basename(URL))
download.file(URL, tmp)
piggyback::pb_upload(file = tmp,
                     repo = "RajLabMSSM/catalogueR")
}
}
\usage{
COLOC_get_example_res(save_dir = tempdir(), full = FALSE)
}
\arguments{
\item{save_dir}{Local directory to cache data in.}

\item{full}{Download the non-filtered version of the colocalization results
(Default: \code{FALSE}).}
}
\description{
Example colocalization results from running \link[catalogueR]{COLOC_run}.
on GWAS summary stats from all loci in
\href{https://doi.org/10.1016/S1474-4422(19)30320-5}{Nalls23andMe_2019}.
These results were published in
\href{https://pubmed.ncbi.nlm.nih.gov/34617105/}{
Schilder & Raj (Human Molecular Genetics, 2021)}
}
\examples{
coloc_QTLs <- catalogueR::COLOC_get_example_res()
}
\seealso{
Other coloc: 
\code{\link{COLOC_corplot}()},
\code{\link{COLOC_get_res}()},
\code{\link{COLOC_heatmap}()},
\code{\link{COLOC_merge_res}()},
\code{\link{COLOC_report_summary}()},
\code{\link{COLOC_run}()}
}
\concept{coloc}
