% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC_heatmap.R
\name{COLOC_heatmap}
\alias{COLOC_heatmap}
\title{Plot summary of coloc tests: heatmap}
\usage{
COLOC_heatmap(
  coloc_QTLs,
  coloc_thresh = 0.8,
  label_snp_groups = FALSE,
  save_dir = tempdir(),
  show_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{coloc_QTLs}{Colocalization results from \link[catalogueR]{COLOC_run}.}

\item{coloc_thresh}{Colocalization Posterior Probability threshold,
 using the formula:
\code{(PP.H3 + PP.H4 >= coloc_thresh) & (PP.H4 / PP.H3 >= 2}.}

\item{label_snp_groups}{Whether to label SNP groups.}

\item{save_dir}{Directory to save plot in.}

\item{show_plot}{Whether to print the plot.}

\item{verbose}{Print messages.}
}
\description{
Plots the output from \link[catalogueR]{COLOC_run}.
}
\examples{
coloc_QTLs <- catalogueR::COLOC_get_example_res()
gg_coloc <- catalogueR::COLOC_heatmap(
    coloc_QTLs = coloc_QTLs,
    coloc_thresh = .5)
}
\seealso{
Other coloc: 
\code{\link{COLOC_corplot}()},
\code{\link{COLOC_get_example_res}()},
\code{\link{COLOC_get_res}()},
\code{\link{COLOC_merge_res}()},
\code{\link{COLOC_report_summary}()},
\code{\link{COLOC_run}()}
}
\concept{coloc}
