% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC_plot_loci.R
\name{COLOC_plot_loci}
\alias{COLOC_plot_loci}
\title{Faceted Manhattan plots of QTL datasets}
\usage{
COLOC_plot_loci(
  gwas.qtl_paths = NULL,
  coloc_QTLs = NULL,
  plot_dat = NULL,
  qtl_thresh = 1e-05,
  coloc_thresh = 0.8,
  gwas_label = "GWAS",
  remove_extra_panes = TRUE,
  y_facet_angle = 0,
  x_facet_angle = 270,
  show_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gwas.qtl_paths}{Query results paths from 
\link[catalogueR]{eQTLcatalogue_query}.}

\item{coloc_QTLs}{Colocalization results from \link[catalogueR]{COLOC_run}.}

\item{plot_dat}{[Optional] Pre-computed plot data from
\link[catalogueR]{COLOC_merge_res}.}

\item{qtl_thresh}{QTL uncorrected p-value ("pvalues.QTL") threshold.}

\item{coloc_thresh}{Colocalization Posterior Probability threshold,
 using the formula:
\code{(PP.H3 + PP.H4 >= coloc_thresh) & (PP.H4 / PP.H3 >= 2}.}

\item{gwas_label}{Label for the GWAS subplot.}

\item{remove_extra_panes}{Remove SNPs from non-significant panes.}

\item{y_facet_angle}{Angle of the y-axis facet labels.}

\item{x_facet_angle}{Angle of the x-axis facet labels.}

\item{show_plot}{Whether to print the plot.}

\item{verbose}{Print messages.}
}
\description{
Use the coloc results (\code{coloc_QTLs}) to determine which
full summary statistics (\code{gwas.qtl_paths}) to plot.
}
\examples{
coloc_QTLs <- catalogueR::COLOC_get_example_res()
gwas.qtl_paths <- catalogueR::eQTLcatalogue_example_queries()
gg_gwas.qtl <- catalogueR::COLOC_plot_loci(
    gwas.qtl_paths = gwas.qtl_paths,
    coloc_QTLs = coloc_QTLs,
    coloc_thresh = .5,
    qtl_thresh = .005,
    remove_extra_panes = FALSE)  
}
\concept{plots}
