% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coloc_QTLs}
\alias{coloc_QTLs}
\title{Example colocalization results}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 155577 rows and 21 columns.
}
\usage{
coloc_QTLs
}
\description{
Example colocalization results from running \code{\link{catalogueR::run_coloc}}
on GWAS summary stats from all loci in \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls23andMe_2019}.
}
\examples{
\dontrun{
library(dplyr)
root_dir <- "~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019"
sub_dir <- "_genome_wide/COLOC/coloc.eQTL_Catalogue_ALL.csv.gz"
coloc_QTLs <- data.table::fread(file.path(root_dir, sub_dir))
coloc_QTLs <- coloc_QTLs \%>\% dplyr::rename(gene.QTL=eGene, qtl_id=qtl.id)
usethis::use_data(coloc_QTLs, overwrite=T)
}
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{get_colocs}()},
\code{\link{plot_coloc_summary}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
\keyword{datasets}
