% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensembl_to_hgnc}
\alias{ensembl_to_hgnc}
\title{Convert ENSEMBL IDs to HGNC gene symbols}
\usage{
ensembl_to_hgnc(ensembl_ids, unique_only = T, verbose = T)
}
\description{
Convert ENSEMBL IDs to HGNC gene symbols
}
\examples{
ensembl_ids <- c("ENSG00000176697","ENSG00000128573","ENSG00000109743")
gene_symbols <- ensembl_to_hgnc(ensembl_ids=ensembl_ids)
}
\seealso{
Other utils: 
\code{\link{gather_files}()},
\code{\link{hgnc_to_ensembl}()},
\code{\link{liftover}()}
}
\concept{utils}
