% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{eQTL_Catalogue.annotate_tissues}
\alias{eQTL_Catalogue.annotate_tissues}
\title{Annotate QTL datasets with metadata}
\usage{
eQTL_Catalogue.annotate_tissues(dat, add_tissue_counts = F)
}
\description{
Annotate QTL datasets with metadata
}
\examples{
dat <- eQTL_Catalogue.annotate_tissues(dat)
}
\keyword{metadata}
