% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{eQTL_Catalogue.fetch}
\alias{eQTL_Catalogue.fetch}
\title{2. Query eQTL Catalogue datasets by region}
\usage{
eQTL_Catalogue.fetch(
  unique_id,
  quant_method = "ge",
  infer_region = T,
  gwas_data = NULL,
  is_gwas = F,
  nThread = 1,
  use_tabix = T,
  chrom = NULL,
  bp_lower = NULL,
  bp_upper = NULL,
  multithread_tabix = F,
  add_qtl_id = T,
  convert_genes = T,
  conda_env = "echoR",
  verbose = T
)
}
\arguments{
\item{quant_method}{eQTL Catalogue actually contains more than just eQTL data.
For each dataset, the following kinds of QTLs can be queried:  
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default}) or \code{quant_method="microarray"}, depending on the dataset. \strong{catalogueR} will automatically select whichever option is available.}  
\item{exon expression QTL}{\emph{*under construction*}  \code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}  \code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{\emph{*under construction*}  \code{quant_method="txrev"}}   
}}

\item{nThread}{The number of CPU cores you want to use to speed up your queries through parallelization.}

\item{use_tabix}{Tabix is about ~17x faster (\emph{default:} =T) than the REST API (\emph{=F}).}

\item{bp_lower}{Minimum basepair position of the query window.}

\item{bp_upper}{Maxmimum basepair position of the query window.}

\item{verbose}{Show more (\code{=T}) or fewer (\code{=F}) messages.}
}
\description{
Choose between tabix (faster for larger queries) 
or RESTful API (faster for small queries).
}
\examples{
data("meta"); data("BST1");
gwas.qtl <- eQTL_Catalogue.fetch(unique_id=meta$unique_id[1], gwas_data=BST1)
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{LRRK2__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.header}},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
