% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eQTL_Catalogue.header}
\alias{eQTL_Catalogue.header}
\title{eQTL Catalogue tabix header}
\format{
An object of class \code{character} of length 19.
}
\usage{
eQTL_Catalogue.header
}
\description{
eQTL Catalogue tabix header
}
\examples{
\dontrun{
eQTL_Catalogue.header <- tabix_header(force_new_header = T)
usethis::use_data(eQTL_Catalogue.header, overwrite=T)
}
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{LRRK2__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{datasets}
