% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_Catalogue.list_datasets.R
\name{eQTL_Catalogue.list_datasets}
\alias{eQTL_Catalogue.list_datasets}
\title{List available eQTL datasets}
\usage{
eQTL_Catalogue.list_datasets(save_dir = F, force_new = F, verbose = FALSE)
}
\description{
Does some additional preprocessing of metadata to categorize tissue types.
}
\examples{
meta <- eQTL_Catalogue.list_datasets()
}
\keyword{metadata}
