% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLcatalogue_example_queries.R
\name{eQTLcatalogue_example_queries}
\alias{eQTLcatalogue_example_queries}
\title{Paths to example eQTL Catalogue query results}
\source{
\url{https://doi.org/10.1016/S1474-4422(19)30320-5}

\code{
paths <- catalogueR:::eQTLcatalogue_example_queries_local()
fnames <- lapply(paths, function(x){
    piggyback::pb_upload(file = x, 
                         repo = "RajLabMSSM/catalogueR")
    return(basename(x))
})
}
}
\usage{
eQTLcatalogue_example_queries(
  save_dir = tempdir(),
  fnames = c("BST1__Alasoo_2018.macrophage_IFNg+Salmonella.tsv.gz",
    "BST1__Alasoo_2018.macrophage_IFNg.tsv.gz",
    "BST1__Alasoo_2018.macrophage_naive.tsv.gz",
    "BST1__Alasoo_2018.macrophage_Salmonella.tsv.gz",
    "LRRK2__Alasoo_2018.macrophage_IFNg.tsv.gz",
    "LRRK2__Alasoo_2018.macrophage_naive.tsv.gz",
    "MEX3C__Alasoo_2018.macrophage_IFNg.tsv.gz",
    "MEX3C__Alasoo_2018.macrophage_naive.tsv.gz")
)
}
\arguments{
\item{save_dir}{Local directory to cache data in.}

\item{fnames}{Character vector of file names to download.}
}
\value{
Path to merged example GWAS-QTL summary statistics.
}
\description{
Returns the paths to eQTL Catalogue query results stored within
\pkg{catalogueR}. Each file is a merged \link[data.table]{data.table}
 of the GWAS summary stats used to make
the query, and the eQTL Catalogue query results
(which can contain data for multiple eGenes).
}
\details{
GWAS data originally comes from the Parkinson's disease GWAS
by \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al., 2019 (The Lancet Neurology)}.
}
\examples{
gwas.qtl_paths <- catalogueR::eQTLcatalogue_example_queries()
}
\concept{Nalls23andMe_2019}
