% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLcatalogue_search_metadata.R
\name{eQTLcatalogue_search_metadata}
\alias{eQTLcatalogue_search_metadata}
\title{Search eQTL Catalogue metadata}
\usage{
eQTLcatalogue_search_metadata(qtl_search = NULL, verbose = TRUE)
}
\arguments{
\item{qtl_search}{This function will automatically search for any datasets
that match your criterion.
For example, if you search "Alasoo_2018", it will query the datasets:
\itemize{
\item{Alasoo_2018.macrophage_naive}
\item{Alasoo_2018.macrophage_Salmonella}
\item{Alasoo_2018.macrophage_IFNg+Salmonella}
}
You can be more specific about which datasets you want to include,
for example by searching: "Alasoo_2018.macrophage_IFNg".
You can even search by tissue or condition type
(e.g. \code{c("blood","brain")})
and any QTL datasets containing those substrings (case-insensitive)
in their name or metadata will be queried too.}

\item{verbose}{Show more (\code{=TRUE}) or fewer (\code{=FALSE}) messages.}
}
\description{
Searches through multiple relevant metadata columns to find
eQTL Catalogue datasets
that match at least one of your substrings in a list.
All searches are case-insensitive.
If \code{qtl_search=NULL}, will return all available datasets.
}
\examples{
qtl_datasets <- eQTLcatalogue_search_metadata(qtl_search = c(
    "Alasoo_2018",
    "monocyte"
))
qtl_datasets.brain <- eQTLcatalogue_search_metadata(qtl_search = "brain")
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTLcatalogue_fetch}()},
\code{\link{eQTLcatalogue_header}},
\code{\link{eQTLcatalogue_iterate_fetch}()},
\code{\link{eQTLcatalogue_query}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
