% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_restAPI.R
\name{fetch_restAPI}
\alias{fetch_restAPI}
\title{2. Query eQTL Catalogue datasets by region}
\source{
\code{
data("meta")
query_granges <- echotabix::construct_query(query_dat = echodata::BST1)
qtl.subset <- fetch_restAPI(dataset_id = meta$dataset_id[1],
                            query_granges = query_granges)
}
}
\usage{
fetch_restAPI(
  dataset_id,
  query_granges,
  quant_method = "ge",
  is_gwas = FALSE,
  size = NULL,
  timeout = 5 * 60,
  verbose = TRUE
)
}
\arguments{
\item{quant_method}{eQTL Catalogue actually contains more than just
 eQTL data.
For each dataset, the following kinds of QTLs can be queried:
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default})
or \code{quant_method="microarray"}, depending on the dataset.
\strong{catalogueR} will automatically select whichever option is available.}
\item{exon expression QTL}{\emph{*under construction*}
\code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}
\code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{
\emph{*under construction*}  \code{quant_method="txrev"}}
}}

\item{verbose}{Show more (\code{=TRUE}) or fewer (\code{=FALSE}) messages.}
}
\description{
2.2 Method 2: RESTful API
Slower than tabix (unless you're only querying several specific SNPs).
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTLcatalogue_fetch}()},
\code{\link{eQTLcatalogue_header}},
\code{\link{eQTLcatalogue_iterate_fetch}()},
\code{\link{eQTLcatalogue_query}()},
\code{\link{eQTLcatalogue_search_metadata}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{internal}
