% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_restAPI.R
\name{fetch_restAPI}
\alias{fetch_restAPI}
\title{2. Query eQTL Catalogue datasets by region}
\source{
\code{
data("meta")
gwas_data <- echodata::BST1 
qtl.subset <- fetch_restAPI(unique_id = meta$unique_id[1],
                            gwas_data = gwas_data)
}
}
\usage{
fetch_restAPI(
  unique_id,
  quant_method = "ge",
  infer_region = TRUE,
  gwas_data = NULL,
  chrom = NULL,
  bp_lower = NULL,
  bp_upper = NULL,
  is_gwas = FALSE,
  size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{quant_method}{eQTL Catalogue actually contains more than just
 eQTL data.
For each dataset, the following kinds of QTLs can be queried:
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default}) 
or \code{quant_method="microarray"}, depending on the dataset. 
\strong{catalogueR} will automatically select whichever option is available.}
\item{exon expression QTL}{\emph{*under construction*}  
\code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}  
\code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{
\emph{*under construction*}  \code{quant_method="txrev"}}
}}

\item{verbose}{Show more (\code{=TRUE}) or fewer (\code{=FALSE}) messages.}
}
\description{
2.2 Method 2: RESTful API
Slower than tabix (unless you're only querying several specific SNPs).
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{LRRK2__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.header}},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_tabix}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
\keyword{internal}
