% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fetch_tabix}
\alias{fetch_tabix}
\title{2. Query eQTL Catalogue datasets by region}
\usage{
fetch_tabix(
  unique_id,
  quant_method = "ge",
  infer_region = T,
  gwas_data = NULL,
  chrom = NULL,
  bp_lower = NULL,
  bp_upper = NULL,
  is_gwas = F,
  nThread = 4,
  verbose = T
)
}
\arguments{
\item{quant_method}{eQTL Catalogue actually contains more than just eQTL data.
For each dataset, the following kinds of QTLs can be queried:  
\describe{
\item{gene expression QTL}{\code{quant_method="ge"} (\emph{default}) or \code{quant_method="microarray"}, depending on the dataset. \strong{catalogueR} will automatically select whichever option is available.}  
\item{exon expression QTL}{\emph{*under construction*}  \code{quant_method="ex"}}
\item{transcript usage QTL}{\emph{*under construction*}  \code{quant_method="tx"}}
\item{promoter, splice junction and 3' end usage QTL}{\emph{*under construction*}  \code{quant_method="txrev"}}   
}}

\item{nThread}{The number of CPU cores you want to use to speed up your queries through parallelization.}

\item{verbose}{Show more (\code{=T}) or fewer (\code{=F}) messages.}
}
\description{
2.1 Method 1: Tabix
Faster alternative to REST API.
}
\examples{
data("meta"); data("BST1");
qtl.subset <- fetch_tabix(unique_id=meta$unique_id[1], gwas_dat=BST1)
}
\seealso{
Other eQTL Catalogue: 
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{merge_gwas_qtl}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
