% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{get_colocs}
\alias{get_colocs}
\title{Run coloc on GWAS-QTL object}
\usage{
get_colocs(
  qtl.egene,
  gwas.region,
  merge_by_rsid = T,
  PP_threshold = 0.8,
  verbose = T
)
}
\description{
Run coloc on GWAS-QTL object
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
