% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_size.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Infer sample size}
\source{
\code{
subset_DT <- echodata::BST1
subset_DT <- catalogueR:::get_sample_size(subset_DT = subset_DT)
}
}
\usage{
get_sample_size(
  subset_DT,
  sample_size = NULL,
  effective_ss = TRUE,
  verbose = TRUE
)
}
\description{
Detect or create column with (effective) sample size from 
summary statistics data.
}
\concept{general}
\keyword{internal}
