% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_to_ensembl.R
\name{hgnc_to_ensembl}
\alias{hgnc_to_ensembl}
\title{Convert HGNC gene symbols to ENSEMBL IDs}
\usage{
hgnc_to_ensembl(gene_symbols, unique_only = TRUE, verbose = TRUE)
}
\description{
Convert HGNC gene symbols to ENSEMBL IDs
}
\examples{
gene_symbols <- c("BDNF", "FOXP2", "BST1")
ensembl_ids <- hgnc_to_ensembl(gene_symbols)
}
\seealso{
Other utils: 
\code{\link{ensembl_to_hgnc}()},
\code{\link{gather_files}()},
\code{\link{liftover}()}
}
\concept{utils}
