% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_plot.R
\name{locus_plot}
\alias{locus_plot}
\title{Locus plot}
\usage{
locus_plot(
  coloc_QTLs,
  locus_name,
  LD_reference = "1KGphase3",
  LD_results_dir = file.path(tempdir(), "plots_LD"),
  plot_results_dir = file.path(tempdir(), "plots"),
  force_new_LD = FALSE,
  plot_zoom = "1x",
  QTL_title = "eQTL Catalogue",
  GWAS_title = "GWAS",
  GWAS_x_lab = NULL,
  show_plot = TRUE,
  save_plot = TRUE,
  dpi = 350,
  height = 12,
  width = 7
)
}
\arguments{
\item{coloc_QTLs}{Colocalization results from \link[catalogueR]{run_coloc}.}

\item{locus_name}{Locus name.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.}
}}

\item{LD_results_dir}{Directory to store LD matrices in.}

\item{plot_results_dir}{Directory to store plots in.}

\item{force_new_LD}{If LD file exists, create a new one.}

\item{plot_zoom}{How many times to zoom into plot.}

\item{QTL_title}{QTL subplot title.}

\item{GWAS_title}{GWAS subplot title.}

\item{GWAS_x_lab}{GWAS subplot x-axis label.}

\item{show_plot}{Whether to print the plot.}

\item{save_plot}{Whether to save the plot.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{height}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{width}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}
}
\description{
Plot colocalized GWAS and QTL results.
}
