% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Merge files from a list of paths}
\usage{
merge_files(file_paths, nThread = 1, verbose = TRUE)
}
\arguments{
\item{file_paths}{Paths of files to import and merge into one 
\link[data.table]{data.table}.}

\item{nThread}{Number of threads to parallelize reading in files across.}

\item{verbose}{Print messages.}
}
\description{
Merge a list of files into one by stacking them on top of each other
 (i.e. \code{rbind}).
}
\examples{
sumstats_paths <- echodata:: get_Nalls2019_loci(limit_snps = 5)
merged_dat <- merge_files(file_paths = sumstats_paths)
}
\seealso{
Other utils: 
\code{\link{ensembl_to_hgnc}()},
\code{\link{hgnc_to_ensembl}()}
}
\concept{utils}
