% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{merge_gwas_qtl}
\alias{merge_gwas_qtl}
\title{Merge GWAS data (query) and QTL data (results)}
\usage{
merge_gwas_qtl(gwas_data, qtl.subset, verbose = T)
}
\arguments{
\item{verbose}{Show more (\code{=T}) or fewer (\code{=F}) messages.}
}
\description{
Merge GWAS data (query) and QTL data (results)
}
\seealso{
Other eQTL Catalogue: 
\code{\link{BST1__Alasoo_2018.macrophage_IFNg}},
\code{\link{LRRK2__Alasoo_2018.macrophage_IFNg}},
\code{\link{MEX3C__Alasoo_2018.macrophage_IFNg}},
\code{\link{eQTL_Catalogue.fetch}()},
\code{\link{eQTL_Catalogue.header}},
\code{\link{eQTL_Catalogue.iterate_fetch}()},
\code{\link{eQTL_Catalogue.query}()},
\code{\link{eQTL_Catalogue.search_metadata}()},
\code{\link{fetch_restAPI}()},
\code{\link{fetch_tabix}()},
\code{\link{meta}}
}
\concept{eQTL Catalogue}
