% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_locus_plot.R
\name{multi_locus_plot}
\alias{multi_locus_plot}
\title{Faceted Manhattan plots of QTL datasets}
\usage{
multi_locus_plot(
  gwas.qtl_paths = NULL,
  coloc_QTLs = NULL,
  plot_dat = NULL,
  qtl_thresh = 1e-05,
  coloc_thresh = 0.8,
  gwas_label = "GWAS",
  remove_extra_panes = TRUE,
  y_facet_angle = 0,
  x_facet_angle = 270,
  show_plot = TRUE,
  verbose = TRUE
)
}
\description{
Use the coloc results (coloc_QTLs) to determine which full summary stats (gwas.qtl_paths) to plot.
}
\examples{
data("coloc_QTLs")
gwas.qtl_paths <- example_eQTL_Catalogue_query_paths()
gg_gwas.qtl <- multi_locus_plot(gwas.qtl_paths = gwas.qtl_paths, coloc_QTLs = coloc_QTLs, coloc_thresh = .5, qtl_thresh = .005, remove_extra_panes = FALSE)
\dontrun{
data("coloc_QTLs_full")
library(ggplot2)
library(dplyr)
root_dir <- "/pd-omics/brian/catalogueR_queries/Nalls23andMe_2019"
gwas.qtl_paths <- list.files(root_dir, pattern = "*.tsv.gz", recursive = TRUE, full.names = TRUE)
gwas.qtl_paths <- gsub("___", "__", gwas.qtl_paths)
colnames(coloc_QTLs_full) <- gsub("\\\\.df1", ".QTL", colnames(coloc_QTLs_full))
colnames(coloc_QTLs_full) <- gsub("\\\\.df2", ".GWAS", colnames(coloc_QTLs_full))
plot_dat <- gather_colocalized_data(gwas.qtl_paths = gwas.qtl_paths, coloc_QTLs = coloc_QTLs_full, qtl_thresh = NULL, coloc_thresh = .95)
gg_gwas.qtl <- multi_locus_plot(plot_dat = plot_dat)
}
}
\concept{plots}
