% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coloc_summary.R
\name{plot_coloc_summary}
\alias{plot_coloc_summary}
\title{Plot summary of coloc tests}
\usage{
plot_coloc_summary(
  coloc_QTLs,
  PP_thresh = 0.8,
  label_snp_groups = F,
  save_dir = ".",
  show_plot = TRUE
)
}
\description{
Plots the output from \code{catalogueR::run_coloc}.
}
\examples{
data("coloc_QTLs")
gg_coloc <- plot_coloc_summary(coloc_QTLs = coloc_QTLs, PP_thresh = .99, label_snp_groups = FALSE)
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{coloc_QTLs_full}},
\code{\link{coloc_QTLs}},
\code{\link{gather_colocalized_data}()},
\code{\link{get_colocs}()},
\code{\link{retrieve_sumstats_info}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
