% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{plot_coloc_summary}
\alias{plot_coloc_summary}
\title{Plot summary of coloc tests}
\usage{
plot_coloc_summary(
  coloc_QTLs,
  PP_thresh = 0.8,
  label_snp_groups = F,
  save_dir = ".",
  show_plot = T
)
}
\description{
Plots the output from \code{catalogueR::run_coloc}.
}
\examples{
data("coloc_QTLs")
gg_coloc <- plot_coloc_summary(coloc_QTLs=coloc_QTLs, PP_thresh=.99, label_snp_groups=F)
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{coloc_QTLs}},
\code{\link{get_colocs}()},
\code{\link{run_coloc}()}
}
\concept{coloc}
