% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{run_coloc}
\alias{run_coloc}
\title{Iteratively run coloc on GWAS-QTL objects}
\usage{
run_coloc(
  gwas.qtl_paths,
  save_path = "./coloc_results.tsv.gz",
  nThread = 3,
  top_snp_only = T,
  split_by_group = F
)
}
\description{
Iteratively run coloc on GWAS-QTL objects
}
\seealso{
Other coloc: 
\code{\link{COLOC.report_summary}()},
\code{\link{get_colocs}()}
}
\concept{coloc}
